chore: update demos and exports for MoveToSquin split
- demo/move_lang.py: use MoveToSquinPhysical
- demo/community_call_jan_28_2026.py: use MoveToSquinPhysical
- demo/pipeline_details.py: use MoveToSquinLogical with generate_logical_noise_model
- __init__.py: export generate_logical_noise_model
