+++ demo/community_call_jan_28_2026.py
@@ -101,10 +101,10 @@ transversal_main.print()
 debugger(transversal_main, arch_spec=physical_arch)
 
 # %%
-from bloqade.lanes import generate_simple_noise_model  # noqa F402
-from bloqade.lanes.transform import MoveToSquin  # noqa F402
+from bloqade.lanes.noise_model import generate_simple_noise_model  # noqa F402
+from bloqade.lanes.transform import MoveToSquinPhysical  # noqa F402
 
-squin_kernel = MoveToSquin(
+squin_kernel = MoveToSquinPhysical(
     physical_arch, noise_model=generate_simple_noise_model()
 ).emit(transversal_main)
 squin_kernel.print()
